+++ android/snapshot-helper/src/main/java/com/callstack/agentdevice/snapshothelper/AccessibilityTreeCapture.java
@@ -69,6 +69,10 @@ final class AccessibilityTreeCapture {
     // destination inside one AndroidComposeView. A persistent helper session reuses a single
     // UiAutomation connection across captures, and its per-connection accessibility node cache is
     // not always invalidated by such a swap. Drop the cache before each traversal.
+    //
+    // Declared residue (agent-device #1832): the reset mechanism differs by API level — the public
+    // clearCache() on 34+, a setServiceInfo() re-apply below — and the two are not known to be
+    // equivalent. This is an acquisition-freshness difference the payload does not disclose.
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
       try {
         automation.clearCache();
