+++ src/daemon/handlers/__tests__/interaction-touch-fixtures.ts
@@ -10,7 +10,7 @@ import { activateCompleteRefFrame } from '../../ref-frame.ts';
 import type { SessionStore } from '../../session-store.ts';
 import type { SessionState } from '../../types.ts';
 import { handleInteractionCommands } from '../interaction.ts';
-import { buildSnapshotState } from '../snapshot-capture.ts';
+import { buildSnapshotState } from '../../snapshot-state.ts';
 
 /**
  * Shared factories for the interaction touch handler tests. Named pure
