+++ src/platforms/android/ui-hierarchy.ts
@@ -2,7 +2,16 @@ import type { RawSnapshotNode, Rect, SnapshotOptions } from '@agent-device/kerne
 import { parseBounds } from '@agent-device/kernel/bounds';
 import { decodeXmlCharacterReferences } from '@agent-device/xml';
 import { isScrollableType } from '@agent-device/contracts/snapshot';
+import {
+  isAgentTarget,
+  isGenericAndroidId,
+  type AndroidNode,
+  type AndroidUiHierarchy,
+} from './ui-hierarchy-node.ts';
+import { collectAndroidHiddenNodes } from './ui-hierarchy-visibility.ts';
 import { scopePresentedAndroidSnapshot } from './ui-hierarchy-scope.ts';
+
+export type { AndroidUiHierarchy } from './ui-hierarchy-node.ts';
 import {
   type AndroidSystemChromeProvenance,
   isAndroidSystemChromeWindowResourceId,
@@ -106,6 +115,8 @@ type AndroidSnapshotBuildState = {
   options: SnapshotOptions;
   analysis: AndroidSnapshotAnalysis;
   interactiveDescendantMemo: Map<AndroidNode, boolean>;
+  /** Subtrees the regular projection hides (invisible / stale window / covered). Empty for raw. */
+  hidden: ReadonlySet<AndroidNode>;
   truncated: boolean;
 };
 
@@ -125,6 +136,9 @@ export function buildUiHierarchySnapshot(
     options,
     analysis: analyzeAndroidTree(tree),
     interactiveDescendantMemo: new Map(),
+    // C3: raw is the acquired tree (normalization only); regular additionally hides what Android
+    // marks invisible, stale application windows, and covered same-window surfaces.
+    hidden: options.raw ? new Set() : collectAndroidHiddenNodes(tree),
     truncated: false,
   };
 
@@ -163,7 +177,7 @@ function walkUiHierarchyNode(
     state.truncated = true;
     return;
   }
-  if (depth > state.maxDepth) return;
+  if (depth > state.maxDepth || state.hidden.has(node)) return;
 
   const include = state.options.raw
     ? true
@@ -221,8 +235,7 @@ function appendAndroidSnapshotNode(
     hittable: isAgentTarget(node) || undefined,
     depth: compactedAndroidNodeDepth(state.nodes, parentIndex),
     parentIndex,
-    ...(node.hiddenContentAbove ? { hiddenContentAbove: true } : {}),
-    ...(node.hiddenContentBelow ? { hiddenContentBelow: true } : {}),
+    ...androidScrollActionHints(node, state.hidden),
     ...(systemChrome ? { systemChrome: true } : {}),
   });
   return currentIndex;
@@ -240,6 +253,7 @@ function hasInteractiveDescendant(state: AndroidSnapshotBuildState, node: Androi
   if (cached !== undefined) return cached;
   for (const child of node.children) {
     if (
+      !state.hidden.has(child) &&
       child.visibleToUser !== false &&
       (isAgentTarget(child) || hasInteractiveDescendant(state, child))
     ) {
@@ -396,40 +410,6 @@ function readXmlAttr(attrs: Map<string, string>, name: string): string | null {
   return attrs.get(name) ?? null;
 }
 
-export type AndroidUiHierarchy = {
-  type: string | null;
-  label: string | null;
-  value: string | null;
-  identifier: string | null;
-  packageName: string | null;
-  rect?: Rect;
-  enabled?: boolean;
-  visibleToUser?: boolean;
-  drawingOrder?: number;
-  focused?: boolean;
-  // Two independent facts, never collapsed, and never undefined: the helper omits false attributes
-  // while stock UiAutomator writes them out, so reading an absent attribute as a value gave two
-  // encodings of one control opposite answers.
-  clickable: boolean;
-  focusable: boolean;
-  depth: number;
-  parentIndex?: number;
-  hiddenContentAbove?: boolean;
-  hiddenContentBelow?: boolean;
-  scrollable?: boolean;
-  canScrollForward?: boolean;
-  canScrollBackward?: boolean;
-  windowIndex?: number;
-  windowType?: number;
-  windowLayer?: number;
-  windowActive?: boolean;
-  windowFocused?: boolean;
-  windowRect?: Rect;
-  children: AndroidNode[];
-};
-
-type AndroidNode = AndroidUiHierarchy;
-
 type AndroidNodeInclusionInfo = {
   type: string;
   hasMeaningfulText: boolean;
@@ -437,26 +417,6 @@ type AndroidNodeInclusionInfo = {
   isStructural: boolean;
   isVisual: boolean;
 };
-
-type AndroidFootprint = {
-  /** Boxes of what the subtree paints: touch targets, scrollables and labelled leaves. */
-  paints: Rect[];
-  /** Boxes of what an agent would see of the subtree: `paints` plus labelled/identified nodes. */
-  shows: Rect[];
-  hasAgentTarget: boolean;
-};
-
-type AndroidTreePruneState = {
-  footprintMemo: WeakMap<AndroidNode, AndroidFootprint>;
-};
-
-type AndroidCoveringCandidate = {
-  node: AndroidNode;
-  drawingOrder: number;
-  footprint: Rect[];
-};
-
-const ANDROID_WINDOW_TYPE_APPLICATION = 1;
 
 export function parseUiHierarchyTree(xml: string): AndroidUiHierarchy {
   const root: AndroidUiHierarchy = {
@@ -514,330 +474,46 @@ export function parseUiHierarchyTree(xml: string): AndroidUiHierarchy {
     }
     match = tokenRegex.exec(xml);
   }
-  // Raw Android snapshots are uncollapsed, but still agent-visible. The helper can expose
-  // aria-hidden/no-hide-descendants children, so prune nodes Android marks hidden to users.
-  pruneAndroidInvisibleSubtrees(root);
-  discardInactiveAndroidApplicationWindows(root);
-  // UiAutomation can expose covered React Native navigation surfaces in the same accessibility
-  // window. If a higher drawing-order sibling covers them, agents should see the foreground surface.
-  pruneAndroidCoveredSubtrees(root, { footprintMemo: new WeakMap() });
-  applyAndroidScrollActionHints(root);
   return root;
 }
 
-/** A node a touch can act on. */
-function isTouchTarget(node: AndroidNode): boolean {
-  return node.clickable;
-}
-
-/** A node D-pad/keyboard traversal can land on. Normal for TV controls, which are rarely clickable. */
-function isFocusTarget(node: AndroidNode): boolean {
-  return node.focusable || node.focused === true;
-}
-
-/** A node an agent can drive by either input model. This is what the public `hittable` projects. */
-function isAgentTarget(node: AndroidNode): boolean {
-  return isTouchTarget(node) || isFocusTarget(node);
-}
-
-/** Text or an address an agent can read or select by. */
-function hasSemanticContent(node: AndroidNode): boolean {
-  return hasMeaningfulLabel(node) || hasMeaningfulIdentifier(node);
-}
-
 /**
- * Focusability is traversal, not paint (#1733), and a label is an announcement, not paint (#1806):
- * a container's content-desc describes its children and an empty labelled View draws nothing. Only
- * a touch target is direct evidence that a node hides what lies under its box.
+ * Scroll-action hints (`hiddenContentAbove/Below`) for the presented node, from the helper's
+ * can-scroll-* attributes. Derived per projection over the children that projection shows: the
+ * overflow estimate that tells a horizontal list from a vertical one must not count children the
+ * regular projection hides, and raw must count them all.
  */
-function hasDirectOcclusionEvidence(node: AndroidNode): boolean {
-  return node.visibleToUser !== false && isTouchTarget(node);
-}
-
-/** Evidence the node is a real surface because it contains something an agent could drive. */
-function hasDescendantOcclusionEvidence(node: AndroidNode, state: AndroidTreePruneState): boolean {
-  return node.children.some(
-    (child) => child.visibleToUser !== false && subtreeFootprint(child, state).hasAgentTarget,
-  );
-}
-
-/**
- * What a subtree paints and what it shows. Paint is the boxes of its touch-consuming surfaces
- * (touch targets, scrollables) and labelled leaves: a full-screen debug overlay
- * holding one floating icon paints only that icon, so it can only hide what sits under the icon,
- * never the whole app behind it (#1806). Rects are kept apart rather than merged into one bounding
- * box: two controls in opposite corners paint two corners, not the screen between them.
- *
- * Shows adds every labelled or identified node — a testID marker or a described container paints
- * nothing, so it never helps a candidate cover, but an agent would still lose it, so it always
- * counts toward what a covered sibling has.
- */
-function subtreeFootprint(node: AndroidNode, state: AndroidTreePruneState): AndroidFootprint {
-  const cached = state.footprintMemo.get(node);
-  if (cached !== undefined) return cached;
-  const footprint = hasPositiveRect(node)
-    ? footprintWithinBox(node, node.rect, state)
-    : childrenFootprint(node, state);
-  state.footprintMemo.set(node, footprint);
-  return footprint;
-}
-
-function footprintWithinBox(
+function androidScrollActionHints(
   node: AndroidNode,
-  ownBox: Rect,
-  state: AndroidTreePruneState,
-): AndroidFootprint {
-  if (paintsOwnBox(node)) {
-    // The whole box is painted; whatever it contains lies inside that box.
-    return { paints: [ownBox], shows: [ownBox], hasAgentTarget: isAgentTarget(node) };
-  }
-  const footprint = childrenFootprint(node, state);
-  if (hasSemanticContent(node)) footprint.shows.push(ownBox);
-  return footprint;
-}
-
-function childrenFootprint(node: AndroidNode, state: AndroidTreePruneState): AndroidFootprint {
-  const footprint: AndroidFootprint = {
-    paints: [],
-    shows: [],
-    hasAgentTarget: isAgentTarget(node),
+  hidden: ReadonlySet<AndroidNode>,
+): { hiddenContentAbove?: true; hiddenContentBelow?: true } {
+  if (!isVerticalScrollableNode(node, hidden)) return {};
+  return {
+    ...(node.canScrollBackward ? { hiddenContentAbove: true as const } : {}),
+    ...(node.canScrollForward ? { hiddenContentBelow: true as const } : {}),
   };
-  for (const child of node.children) {
-    if (child.visibleToUser === false) continue;
-    const childFootprint = subtreeFootprint(child, state);
-    footprint.hasAgentTarget ||= childFootprint.hasAgentTarget;
-    footprint.paints.push(...childFootprint.paints);
-    footprint.shows.push(...childFootprint.shows);
-  }
-  return footprint;
 }
 
-/** Focusability is traversal, not paint (#1733); a container's label describes its children. */
-function paintsOwnBox(node: AndroidNode): boolean {
-  return (
-    isTouchTarget(node) ||
-    node.scrollable === true ||
-    (node.children.length === 0 && hasMeaningfulLabel(node))
-  );
-}
-
-/** Fraction of the covered rects' union that lies under the covering rects' union. */
-function unionCoverage(coveringRects: Rect[], coveredRects: Rect[]): number {
-  const xs = compressedEdges([...coveringRects, ...coveredRects], (rect) => [
-    rect.x,
-    rect.x + rect.width,
-  ]);
-  const ys = compressedEdges([...coveringRects, ...coveredRects], (rect) => [
-    rect.y,
-    rect.y + rect.height,
-  ]);
-  const covering = markCells(coveringRects, xs, ys);
-  const covered = markCells(coveredRects, xs, ys);
-  let coveredArea = 0;
-  let overlapArea = 0;
-  for (let column = 0; column < xs.length - 1; column += 1) {
-    const width = xs[column + 1]! - xs[column]!;
-    for (let row = 0; row < ys.length - 1; row += 1) {
-      const cell = column * (ys.length - 1) + row;
-      if (!covered[cell]) continue;
-      const area = width * (ys[row + 1]! - ys[row]!);
-      coveredArea += area;
-      if (covering[cell]) overlapArea += area;
-    }
-  }
-  return coveredArea <= 0 ? 0 : overlapArea / coveredArea;
-}
-
-function compressedEdges(rects: Rect[], edgesOf: (rect: Rect) => [number, number]): number[] {
-  return [...new Set(rects.flatMap(edgesOf))].sort((left, right) => left - right);
-}
-
-function markCells(rects: Rect[], xs: number[], ys: number[]): Uint8Array {
-  const rows = ys.length - 1;
-  const cells = new Uint8Array((xs.length - 1) * rows);
-  for (const rect of rects) {
-    const firstColumn = xs.indexOf(rect.x);
-    const lastColumn = xs.indexOf(rect.x + rect.width);
-    const firstRow = ys.indexOf(rect.y);
-    const lastRow = ys.indexOf(rect.y + rect.height);
-    for (let column = firstColumn; column < lastColumn; column += 1) {
-      cells.fill(1, column * rows + firstRow, column * rows + lastRow);
-    }
-  }
-  return cells;
-}
-
-/**
- * A childless sibling that only presents: an RN screen-level testID, or a label drawn inside a
- * higher sibling's box (Telegram's `+` over the country-code EditText). Geometry cannot tell a
- * transparent overlay from an opaque one, and exempting a leaf cannot resurrect a covered surface.
- */
-function isPresentationLeaf(node: AndroidNode): boolean {
-  return node.children.length === 0 && !isAgentTarget(node) && hasSemanticContent(node);
-}
-
-function pruneAndroidInvisibleSubtrees(node: AndroidNode): void {
-  let keptCount = 0;
-  for (const child of node.children) {
-    if (child.visibleToUser === false) continue;
-    pruneAndroidInvisibleSubtrees(child);
-    node.children[keptCount] = child;
-    keptCount += 1;
-  }
-  if (keptCount < node.children.length) {
-    node.children.length = keptCount;
-  }
-}
-
-function pruneAndroidCoveredSubtrees(node: AndroidNode, state: AndroidTreePruneState): void {
-  for (const child of node.children) {
-    pruneAndroidCoveredSubtrees(child, state);
-  }
-  if (node.children.length < 2) {
-    return;
-  }
-  const siblings = node.children;
-  const coveringCandidates = siblings
-    .map((sibling) => coveringCandidateOf(sibling, state))
-    .filter((candidate) => candidate !== null);
-  if (coveringCandidates.length === 0) return;
-  node.children = siblings.filter((child) =>
-    shouldKeepAndroidSibling(child, coveringCandidates, state),
-  );
-}
-
-function shouldKeepAndroidSibling(
-  node: AndroidNode,
-  coveringCandidates: AndroidCoveringCandidate[],
-  state: AndroidTreePruneState,
-): boolean {
-  return (
-    isPresentationLeaf(node) ||
-    !isCoveredByHigherDrawingOrderSibling(node, coveringCandidates, state)
-  );
-}
-
-/**
- * Covered means everything an agent would see of the sibling lies under what the candidate paints,
- * by actual overlapped area. Comparing footprints rather than boxes lets two stacked screens with the
- * same layout margins still register as covered, while a sparse overlay never condemns a rich
- * screen however far apart its controls sit.
- */
-function isCoveredByHigherDrawingOrderSibling(
-  node: AndroidNode,
-  coveringCandidates: AndroidCoveringCandidate[],
-  state: AndroidTreePruneState,
-): boolean {
-  if (node.visibleToUser === false || node.drawingOrder === undefined || !hasPositiveRect(node)) {
-    return false;
-  }
-  const shows = subtreeFootprint(node, state).shows;
-  const coveredRects = shows.length > 0 ? shows : [node.rect];
-  for (const candidate of coveringCandidates) {
-    if (candidate.node === node || candidate.drawingOrder <= node.drawingOrder) {
-      continue;
-    }
-    if (unionCoverage(candidate.footprint, coveredRects) >= 0.9) {
-      return true;
-    }
-  }
-  return false;
-}
-
-function hasMeaningfulIdentifier(node: AndroidNode): boolean {
-  const identifier = node.identifier?.trim() ?? '';
-  return Boolean(identifier && !isGenericAndroidId(identifier));
-}
-
-/** The single occlusion classification. Covering is never re-derived from a raw attribute. */
-function coveringCandidateOf(
-  node: AndroidNode,
-  state: AndroidTreePruneState,
-): AndroidCoveringCandidate | null {
-  const { drawingOrder } = node;
-  if (node.visibleToUser === false || drawingOrder === undefined || !hasPositiveRect(node)) {
-    return null;
-  }
-  if (!hasDirectOcclusionEvidence(node) && !hasDescendantOcclusionEvidence(node, state)) {
-    return null;
-  }
-  const footprint = subtreeFootprint(node, state).paints;
-  return footprint.length > 0 ? { node, drawingOrder, footprint } : null;
-}
-
-function hasMeaningfulLabel(node: AndroidNode): boolean {
-  const label = node.label?.trim() ?? '';
-  return Boolean(label && !isGenericAndroidId(label));
-}
-
-function hasPositiveRect(node: AndroidNode): node is AndroidNode & { rect: Rect } {
-  return Boolean(node.rect && node.rect.width > 0 && node.rect.height > 0);
-}
-
-function applyAndroidScrollActionHints(root: AndroidUiHierarchy): void {
-  const stack = [...root.children];
-  while (stack.length > 0) {
-    const node = stack.pop() as AndroidNode;
-    stack.push(...node.children);
-    if (!isVerticalScrollableNode(node)) continue;
-    if (node.canScrollBackward) node.hiddenContentAbove = true;
-    if (node.canScrollForward) node.hiddenContentBelow = true;
-  }
-}
-
-function discardInactiveAndroidApplicationWindows(root: AndroidUiHierarchy): void {
-  const windows = root.children.filter(isAndroidWindowRoot);
-  if (windows.length < 2) return;
-
-  // Android can keep stale application windows in the accessibility tree after drawer and
-  // navigation transitions. Keep dialogs/system windows, but expose only the foreground
-  // application layer so agents do not act on content that is hidden from users.
-  const foregroundApplicationWindows = windows.filter(
-    (window) => isAndroidApplicationWindow(window) && isAndroidForegroundWindow(window),
-  );
-  if (foregroundApplicationWindows.length === 0) return;
-  const foregroundLayer = highestAndroidWindowLayer(foregroundApplicationWindows);
-
-  root.children = root.children.filter((window) => {
-    if (!isAndroidApplicationWindow(window)) return true;
-    if (!isAndroidForegroundWindow(window)) return false;
-    return foregroundLayer === undefined || window.windowLayer === foregroundLayer;
-  });
-}
-
-function highestAndroidWindowLayer(windows: AndroidNode[]): number | undefined {
-  const layers = windows
-    .map((window) => window.windowLayer)
-    .filter((layer): layer is number => layer !== undefined);
-  return layers.length > 0 ? Math.max(...layers) : undefined;
-}
-
-function isAndroidWindowRoot(node: AndroidNode): boolean {
-  return node.windowIndex !== undefined || node.windowType !== undefined;
-}
-
-function isAndroidApplicationWindow(node: AndroidNode): boolean {
-  return node.windowType === ANDROID_WINDOW_TYPE_APPLICATION;
-}
-
-function isAndroidForegroundWindow(node: AndroidNode): boolean {
-  return node.windowActive === true || node.windowFocused === true;
-}
-
-function isVerticalScrollableNode(node: AndroidNode): boolean {
+function isVerticalScrollableNode(node: AndroidNode, hidden: ReadonlySet<AndroidNode>): boolean {
   if (!node.scrollable || !isScrollableType(node.type)) return false;
   const type = `${node.type ?? ''}`.toLowerCase();
   if (type.includes('horizontalscrollview')) return false;
-  const overflow = estimateChildOverflow(node);
+  const overflow = estimateChildOverflow(node, hidden);
   if (overflow && overflow.horizontal > overflow.vertical && overflow.horizontal > 16) {
     return false;
   }
   return true;
 }
 
-function estimateChildOverflow(node: AndroidNode): { horizontal: number; vertical: number } | null {
-  if (!node.rect || node.children.length === 0) return null;
-  const childRects = node.children.map((child) => child.rect).filter((rect) => rect !== undefined);
+function estimateChildOverflow(
+  node: AndroidNode,
+  hidden: ReadonlySet<AndroidNode>,
+): { horizontal: number; vertical: number } | null {
+  const children = node.children.filter(
+    (child) => !hidden.has(child) && child.visibleToUser !== false,
+  );
+  if (!node.rect || children.length === 0) return null;
+  const childRects = children.map((child) => child.rect).filter((rect) => rect !== undefined);
   if (childRects.length === 0) return null;
   const minX = Math.min(...childRects.map((rect) => rect.x));
   const maxX = Math.max(...childRects.map((rect) => rect.x + rect.width));
@@ -952,12 +628,6 @@ function normalizeAndroidType(type: string | null): string {
 function isStructuralAndroidType(type: string): boolean {
   const short = type.split('.').pop() ?? type;
   return short.includes('layout') || short === 'viewgroup' || short === 'view';
-}
-
-function isGenericAndroidId(value: string): boolean {
-  const trimmed = value.trim();
-  if (!trimmed) return false;
-  return /^[\w.]+:id\/[\w.-]+$/i.test(trimmed);
 }
 
 function analyzeAndroidTree(root: AndroidNode): AndroidSnapshotAnalysis {
